requirements.txt
espeakng==1.0.3
openai==1.40.0
PyAudio==0.2.14
pyttsx3==2.90
sounddevice==0.4.7
SpeechRecognition==3.10.4
streamlit==1.37.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
TTS==0.22.0
python-dotenv==1.0.1
